requirements.txt
streamlit
pytz
tzlocal
pycurl
apscheduler
croniter
openpyxl
azure.identity
requests
pandas
